chore: refactor src structure
Hooks & shortcodes are core objects within the system, this
restructuring aims to align them with plugins.

The aim here is to make testing and development easier going
forward as boundaries between parts of the framework have become
more defined.
