+++ src/hooks/index.ts
@@ -1,10 +1,10 @@
 import path from 'path';
 import fs from 'fs-extra';
-import { parseBuildPerf } from './utils';
-import externalHelpers from './externalHelpers';
-import { HookOptions } from './hookInterface/types';
-import prepareShortcodeParser from './utils/prepareShortcodeParser';
-import Page from './utils/Page';
+import { parseBuildPerf } from '../utils';
+import externalHelpers from '../externalHelpers';
+import { HookOptions } from './types';
+import prepareShortcodeParser from '../utils/prepareShortcodeParser';
+import Page from '../utils/Page';
 
 const hooks: Array<HookOptions> = [
   {
