+++ src/index.ts
@@ -4,8 +4,8 @@ export { Elder, build, partialHydration } from './Elder';
 export * from './utils/types';
 export * from './utils/index';
 export * from './routes/routes';
-export * from './hookInterface/types';
-export { hookInterface } from './hookInterface/hookInterface';
-export { hookEntityDefinitions } from './hookInterface/hookEntityDefinitions';
+export * from './hooks/types';
+export { hookInterface } from './hooks/hookInterface';
+export { hookEntityDefinitions } from './hooks/hookEntityDefinitions';
 
 export { default as getElderConfig } from './utils/getConfig';
